Cleaning up ConfigNode.cs
